Folding CampaignListCriteria into CampaignListSegment
Since the segment object holds all of the information needed to
determine eligibility, and the segment object was acting as a factory
for the CampaignListCriteria nodes anyway, I decided to remove the
CampaignListCriteria classes and move the functionality into the
CampaignListSegment classes.  This removes a large set of classes that
were only mirroring the structure of the list segments, anyway.