Error handling and bug fixes
Adding failing test to detect bug in CampaignListMemberMapper where
queried Member__c records were not being queried by "campaign list" id.

Adding detection of malformed Segment__c records in
CampaignListSegmentMapper to detect when Segment__c records are queried
that belong to the same "campaign list", but their parents do not belong
to that same "campaign list".  Now, an exception is thrown when this is
detected.

Adding passing test to verify exception is thrown when malformed
Segment__c record is detected, as noted above.