Modify report row iterator filtering and type
Modifying report row iterator to iterate over ids, specifically, rather
than generic objects.  This is because the iterator now does explicit
comparisons of the values in order to ensure returned ids are unique.

Replacing IteratorFilterUnique with IteratorMonotonicIncreasingIdFilter.
This new iterator filter does away with the set of seen objects (and
associated memory requirements) of IteratorFilterUnique by exploiting
the fact that the ids as returned from the report should be ordered.  By
comparing the next id with the previous id, it can determine whether or
not the id was already returned by checking to see if it is greater than
the previous id.  This works because if the ids from the report are
ordered, but potentially duplicate, then the there are only two cases to
consider for consecutive ids: either the next id is equal to the
previous id, in which case it is a duplicate, or the next id is greater
than the previous id, in which case it is not a duplicate because by
transitivity it will also be greater than every previously returned id.