Modify ReportRowIdIterableIterator to call hasNext() from next()
Previously, next() would call runReport() if it detected that it had
either a) not run the report yet, or b) if it had exhausted the current
page.

However, there was no check to make sure that the next page did indeed
have data to iterate over.

The hasNext() and next() methods have been rewritten so that hasNext()
now runs the report if it detects that it had not yet run the report or
if it has exhausted the current page and the current page indicates that
there is more data to iterate over.  After running the report, the
current page iterator should have data to iterator over if there is a
next value, so hasNext() returns currentPageRowIterator.hasNext().

In next(), it calls hasNext() first, and if that returns false it will
not be able to return anyhting, so it stops executing the method by
returning null.  Otherwise, the call to hasNext() would have set up the
currentPageRowIterator to be on a fresh page (if needed) and so it can
freely call currentPageRowIterator.next().

This should protect against users of the class that call next() without
calling hasNext() first.