Adding grouping node negation
Modify CampaignListSegment to add functionality where grouping segments
(i.e. AndSegment and OrSegments) can be set as negated groupings.  This
required the addition of new members and constructor signatures.  Other
segment class constructor signatures were updated to match the parent
signatures in parameter order.

The big modification here is in the meetsCriteria() method for
AndSegment and OrSegment.  These methods have been rewritten so that
they now effectively ignore grouping segments that do not have any
children.  Unfortunately, this means that some segments may not be able
to return a result when queried for meetsCriteria(), and so this method
may now return null to indicate a result was not able to be computed.