Adding segment persistence and removing old classes
Removing references to CSegment class in Lightning components.  Now,
instead of having the Aura framework serialize CSegment instances, we
pass JSON serialized CampaignListSegment objects to the lightning
component.  The components manipulate these as Javascript objects, and
when saving, pass a JSON serialized version of these objects.

Naming of the fields on this object were simplified to remove references
to CSegment class (i.e., parentCSegment becomes parent, rootCSegment
becomes root, listChildCSegments becomes children). Also removing any
direct references to the Segment__c sObject in the lightning code.

Also removing some other cruft from lightning components, like commented
out code and debug code.

Adding properties to CampaignListSegment classes to allow the type of
the segment to be serialized when the object is serialized.

Add new getSObject() method to CampaignListSegment to be used when
persisting CampaignListSegment objects.  Adding setters for some segment
properties.  These setters will propagate ids down the tree. For
example, if the id of a parent segment is set then this id will be
propagated to that segment's children to update their parentId
references.  Similarly with rootSegmentId.  These ids are only
propagated if they are not null, and care is taken not to set up
circular references where a segment might reference itself as a parent
or root segment.  Also, when a child segment is added, ids from the
parent are also propagated to the children to maintain consistency.

Updating CampaignListSegment to store the name of the source as part of
the object.  This gets populated when querying for a segment tree by
root segment id (by querying the sources at that time).  Formerly, this
name was queried only when getSourceName() was called.  This was changed
to allow the source name to be serialized immediately after querying for
a segment tree.

Updating CampaignList.Segment interface to include setters for segment
properties and new getSObject() method.

Adding new enum SegmentType to be used by CampaignListSegment objects to
keep a reference to their own type for serialization purposes.

Adding upsertSegmentTreeByRootSegment() method to
CampaignListSegmentMapper.  This method will take a segment tree and
persist it to the database as Segment__c objects.  It will update any
existing segments, insert new segments, and remove segments that no
longer exist in that segment tree.

Adding deserializeSegmentFromJSON() method to CampaignListSegmentMapper.
This method will convert a JSON serialized version of a
CampaignListSegment tree and convert it to instances of
CampaignListSegment objects as a similarly formed tree.  In service of
this, there is a class (SegmentData) that exists to deserialize the JSON
data into, and that class will recursively convert to the segments to
the appropriate CampaignListSegment objects.

Updating some licenses that still referenced Salesforce.com Foundation.

Removing classes: CSegment, CampaignPopulate, MembersBuilder,
SegmentMapper, SegmentTreeProcessor, SegmentTreeSave.