Adding report column dropdown to campaign list ui
In order to know which Contact/Lead record to use from a report row, we
must know which column to look at to find the Contact/Lead id.

This adds a dropdown to the UI that will let the user specify the column
when they are defining a report source segment.

Modifying Source.cmp markup to specify CampaignListBuilder_CTRL as a
controller.

Modifying Source.cmp markup to render a ui:inputSelect as an inline form
element whenever rendering the autocomplete input for a report source
segment.

Adding an event handler to Source.cmp markup to handle the change event
whenever the component's segment gets updated.

Adding handleSourceChange() method to SourceController.js that gets
invoked whenever the Source component's segment gets changed.  This
method checks to see if the current segment is a report source segment,
and if so it calls the getReportIdColumns() method on the component's
controller (CampaignListBuilder_CTRL).  If the getReportIdColumns()
method returns a collection of columns, it uses that collection to build
the inputSelect options for the column name component.  If there was an
error for any reason, including the source report doesn't exist, it
populates the column name component with a non-option.  This behavior
should probably be changed as a future enhancement... i.e. handle
different types of errors differently.

Adding handleColumnName() method to SourceController.js to update the
component's segment with the columnName chosen in the column name
inputSelect.  If no column is chosen, we remove the columnName property
from the segment.

Adding getReportIdColumns() AuraEnabled method to
CampaignListBuilter_CTRL.  This method is called by the lightning
component with a report id and delegates to the ReportService to return
id columns in that report.  If the ReportService fails for any reason
(throws an exception), then an exception is thrown in this method which
is checked for in the lightning component as an error status.

Adding getIdColumns() method to ReportService.  This method will
retrieve the metadata for the given report id, and pluck out any columns
that have type 'ID_DATA'.  These are added to a collection and returned.
Note that the columns returned by
ReportExtendedMetadata.getDetailColumnInfo() are actually the columns
available as part of the report type, not necessarily the columns
currently visible in the defined report via the Salesforce UI.

Adding tests for new methods.

Adding ReportService interface.