Fix bug where report columns would not render on first load
The report columns dropdown options were not being populated when
loading a report source where that report source was not the root
segment.  This was because the valueChange event does not get called
when loading these children (for some reason, i dont know why?).  I have
moved the column dropdown populating code to a helper method, and now
delegate to this helper method from the valueChange handler in the
controller.  Additionally, I also make sure to call this helper method
from the renderer, after rendering is first completed.