Rewriting Mutex tests to not reference Campaign fields
The mutex tests were failing in Jenkins because the user under which the
tests were being run was not marked as a 'Marketing User', and the tests
were referencing fields on the Campaign object (since that object is the
intended use case for this class).

Additionally, the tests were referencing a custom field only available
in this repository.

I have rewritten the Mutex tests to use a DateTime field on a standard
object: Contact.EmailBouncedDate.

This field should not require the 'Marketing User' permission, nor
should it require any additional custom fields to be installed for the
Mutex tests to pass.